+++ apps/connect/src/workers/pglite-worker.ts
@@ -1,5 +1,9 @@
-import type { IdbFs } from '@electric-sql/pglite';
-import { type PGliteWorkerOptions, worker } from '@electric-sql/pglite/worker';
+import {
+    type IdbFs,
+    type PGliteWorkerOptions,
+    live,
+    worker,
+} from '@powerhousedao/reactor-browser/pglite';
 
 interface PGLiteWorkerOptions extends PGliteWorkerOptions {
     meta: {
@@ -14,13 +18,18 @@ worker({
             throw new Error('Database name not provided');
         }
 
-        const { IdbFs, PGlite } = await import('@electric-sql/pglite');
+        const { IdbFs, PGlite } = await import(
+            '@powerhousedao/reactor-browser/pglite'
+        );
 
         const idbFs: IdbFs = new IdbFs(databaseName);
         // Create and return a PGlite instance
-        const db = new PGlite({
+        const db = PGlite.create({
             fs: idbFs,
             relaxedDurability: true,
+            extensions: {
+                live,
+            },
         });
 
         return db;