+++ clis/ph-cli/src/services/generate.ts
@@ -1,5 +1,6 @@
 import {
   generate as generateCode,
+  generateDBSchema,
   generateDriveEditor,
   generateEditor,
   generateFromFile,
@@ -8,6 +9,7 @@ import {
   generateSubgraph,
 } from "@powerhousedao/codegen";
 import { getConfig } from "@powerhousedao/config/utils";
+import path from "path";
 
 export type GenerateOptions = {
   interactive?: boolean;
@@ -24,6 +26,8 @@ export type GenerateOptions = {
   importScript?: string;
   file?: string;
   driveEditor?: string;
+  migrationFile?: string;
+  schemaFile?: string;
 };
 
 export async function startGenerate(
@@ -59,6 +63,8 @@ export async function startGenerate(
     importScriptName: options.importScript,
     driveEditor: !!options.driveEditor,
     driveEditorName: options.driveEditor,
+    migrationFile: options.migrationFile,
+    schemaFile: options.schemaFile,
   };
 
   if (command.driveEditor) {
@@ -88,6 +94,12 @@ export async function startGenerate(
     await generateSubgraph(command.subgraphName, options.file || null, config);
   } else if (command.importScript && command.importScriptName) {
     await generateImportScript(command.importScriptName, config);
+  } else if (command.migrationFile) {
+    const { migrationFile, schemaFile } = command;
+    await generateDBSchema({
+      migrationFile: path.join(process.cwd(), migrationFile),
+      schemaFile: schemaFile ? path.join(process.cwd(), schemaFile) : undefined,
+    });
   } else if (filePath) {
     await generateFromFile(filePath, config);
   } else {