+++ packages/codegen/src/codegen/.hygen/templates/powerhouse/generate-processor-operational/migrations.esm.t
@@ -1,0 +1,23 @@
+---
+to: "<%= rootDir %>/<%= h.changeCase.param(name) %>/migrations.ts"
+force: true
+---
+import { type IOperationalStore } from "document-drive/processors/types"
+
+export async function up(db: IOperationalStore): Promise<void> {
+  // Create table 
+  await db.schema
+    .createTable("todo")
+    .addColumn("task", "varchar(255)")
+    .addColumn("status", "boolean")
+    .addPrimaryKeyConstraint("todo_pkey", [
+      "task"
+    ])
+    .ifNotExists()
+    .execute();
+}
+
+export async function down(db: IOperationalStore): Promise<void> {
+  // drop table
+  await db.schema.dropTable("todo").execute();
+}