+++ packages/codegen/src/codegen/.hygen/templates/powerhouse/generate-processor-operational/schema.esm.t
@@ -1,0 +1,5 @@
+---
+to: "<%= rootDir %>/<%= h.changeCase.param(name) %>/schema.ts"
+force: true
+---
+export interface DB {}