+++ packages/reactor-api/src/types.ts
@@ -1,12 +1,14 @@
 import { type IAnalyticsStore } from "@powerhousedao/analytics-engine-core";
-import { type IProcessorManager } from "document-drive/processors/types";
+import {
+  type IOperationalStore,
+  type IProcessorManager,
+} from "document-drive/processors/types";
 import { type Express } from "express";
 import { type GraphQLManager } from "./graphql/graphql-manager.js";
 import { type IPackageManager } from "./packages/types.js";
-import { type Db } from "./utils/db.js";
 
 export type { SubgraphClass } from "./graphql/index.js";
-export type { IPackageLoader, IProcessorHostModule } from "./packages/types.js";
+export type { IPackageLoader } from "./packages/types.js";
 export type { Db } from "./utils/db.js";
 
 export type API = {
@@ -18,5 +20,5 @@ export type API = {
 
 export type ReactorModule = {
   analyticsStore: IAnalyticsStore;
-  operationalStore: Db;
+  operationalStore: IOperationalStore;
 };